updated file name
added runtime-config.ts
